CMR cleanup part 1/?
- Remove old MainContent and Footer
- Rename CMR versions to MainContent and Footer
- Fix feedback link
